project/requirements/requirements_openai.txt
# OpenAI 版本 RAG 系统依赖
# 适用于使用 OpenAI API 的场景

# 核心 LangChain 包
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-openai>=0.0.5

# 向量数据库
chromadb>=0.4.0
sentence-transformers>=2.2.0

# OpenAI 客户端
openai>=1.0.0
python-dotenv>=1.0.0

# 数据处理
numpy>=1.24.0
